+++ lib/lumberjack/formatter/tagged_message.rb
@@ -1,38 +1,15 @@
 # frozen_string_literal: true
 
+require_relative "../message_attributes"
+
 module Lumberjack
-  class Formatter
-    # This class can be used as the return value from a formatter `call` method to
-    # extract additional attributes from an object being logged. This can be useful when there
-    # using structured logging to include important metadata in the log message.
-    #
-    # @example
-    #  # Automatically add attributes with error details when logging an exception.
-    #  logger.add_formatter(Exception, ->(e) {
-    #    Lumberjack::Formatter::TaggedMessage.new(e.inspect, {
-    #      error: {
-    #        message: e.message,
-    #        class: e.class.name,
-    #        stack: e.backtrace
-    #      }
-    #    })
-    #  })
-    class TaggedMessage
-      attr_reader :message, :attributes
-
-      # @param formatter [Formatter] The formatter to apply the transformation to.
-      # @param transform [Proc] The transformation function to apply to the formatted string.
-      def initialize(message, attributes)
-        @message = message
-        @attributes = attributes || {}
-      end
-
-      def to_s
-        inspect
-      end
-
-      def inspect
-        {message: @message, attributes: @attributes}.inspect
+  # This is a deprecated alias for Lumberjack::MessageAttributes.
+  #
+  # @see MessageAttributes
+  class Formatter::TaggedMessage < MessageAttributes
+    def initialize(message, attributes)
+      Utils.deprecated("Lumberjack::Formatter::TaggedMessage", "Use Lumberjack::MessageAttributes instead.") do
+        super
       end
     end
   end